% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{timechange-package}
\alias{timechange}
\alias{timechange-package}
\title{Package \code{timechange}}
\description{
Utilities for efficient updating of date-time object's components while
accounting for time-zones and day-light saving times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timechange/mlvocab/}
  \item Report bugs at \url{https://github.com/vspinu/timechange/issues}
}

}
\author{
Vitalie Spinu (\email{spinuvit@gmail.com})
}
